+++ src/OpenLoco/src/CommandLine.cpp
@@ -4,6 +4,7 @@
 #include "OpenLoco.h"
 #include "S5/S5.h"
 #include "S5/SawyerStream.h"
+#include "Version.h"
 #include <OpenLoco/Core/MemoryStream.h>
 #include <OpenLoco/Diagnostics/Logging.h>
 #include <chrono>
