+++ src/OpenLoco/src/Version.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <string>
+
+namespace OpenLoco
+{
+    std::string getVersionInfo();
+    std::string getPlatformInfo();
+}
